improved sample data creation

+++ database/factories/Bill.php
@@ -171,7 +171,7 @@ $factory->afterCreating(Bill::class, function ($bill, $faker) use ($company) {
             ];
 
             if ($init_status == 'partial') {
-                $payment_request['amount'] = round($amount / 3, $bill->currency->precision);
+                $payment_request['amount'] = (int) round($amount / 3, $bill->currency->precision);
             }
 
             $transaction = dispatch_now(new CreateDocumentTransaction($updated_bill, $payment_request));
